requirements.txt
phidata
streamlit==1.41.1
openai==1.58.1
duckdb==1.1.3
pandas
numpy==2.3.2
agno